Fix (de)serialization of null TIMESTAMP
Use voltdb's NULL representation to represent time.Time{}
over the wire.

The previous implementation used epoch time 0, but this was
inconsistent with csvloader's representation when using
--blank=empty, and has it's obvious down-side.

As time.Time already has a good representation for it's zero
value, checked using Time.IsZero(), this seems to be the
natural and expected behaviour.