Add I/O deadlines.
This adds read/write deadline on every read/write call to Conn.tcpConn,
and also uses the DialTimeout function to bring up the connection.

The new method SetResponseDeadline provide the means of overriding
the default call response deadline.

The tcpConn is now a net.Conn (was *net.TCPConn), as we were only using
the methods from net.Conn and DialTimeout returns this interface.

For your convenience:
* Dial deadline: 5 seconds.
* Login response deadline: 1 second.
* Call response deadline: 2 minutes (as java client default).
* Ping response deadline: 1 second.
* Write deadline: 1 second.